requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
frozenlist==1.4.1
gevent==24.2.1
greenlet==3.0.3
idna==3.7
jmespath==1.0.1
multidict==6.0.5
py_simple_graphql @ file:///D:/projects/GraphQL/py-graphql
pycparser==2.22
python-dotenv==1.0.1
requests==2.32.3
StrEnum==0.4.15
urllib3==2.2.1
websocket-client==1.8.0
websockets==12.0
yarl==1.9.4
zope.event==5.0
zope.interface==6.3
